requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.6.0
anyio==4.1.0
apache-airflow==2.8.0
apache-airflow-providers-common-sql==1.9.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.6.0
apispec==6.3.0
appnope==0.1.4
argcomplete==3.2.1
asgiref==3.7.2
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backoff==2.2.1
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
cron_descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
executing==2.2.0
filelock==3.17.0
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1
googleapis-common-protos==1.62.0
graphviz==0.20.1
greenlet==3.0.2
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.29.1
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.29.5
ipython==9.0.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy-object-proxy==1.10.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
packaging==23.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.3.0
prison==0.2.1
prompt_toolkit==3.0.50
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.2.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==26.2.1
referencing==0.32.0
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.13.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.9.0
tzdata==2025.1
uc-micro-py==1.0.2
unicodecsv==0.14.1
universal-pathlib==0.1.4
urllib3==2.0.6
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.1.1
xgboost==2.1.4
yarl==1.9.4
zipp==3.17.0
